requirements.txt
numpy
ola
protobuf
PyYAML
six
git+https://github.com/allelos/vectors.git
black
pylint
vectors
#xbox360controller
